style/#20 - 위치아이콘 크기위치 수정
